add win screen. add score and power items

+++ core/src/fr/poulpogaz/jam/stage/EnemyAction.java
@@ -2,15 +2,15 @@ package fr.poulpogaz.jam.stage;
 
 public final class EnemyAction<P> {
 
-    private final int start;
+    private final float start;
     private final P pattern;
 
-    public EnemyAction(int start, P pattern) {
+    public EnemyAction(float start, P pattern) {
         this.start = start;
         this.pattern = pattern;
     }
 
-    public int start() {
+    public float start() {
         return start;
     }
